+++ lib/user_agent/gem_version.rb
@@ -2,6 +2,6 @@
 
 class UserAgent
   module GemVersion
-    VERSION = '1.4.0'
+    VERSION = '1.5.0'
   end
 end