+++ spec/browsers/wondery_user_agent_spec.rb
@@ -25,6 +25,23 @@ shared_examples 'Wondery' do |version, platform, os|
   it { expect(useragent).not_to be_bot }
 end
 
+shared_examples 'Wondery Crawler' do |version|
+  it "returns 'Wondery Crawler' as its browser" do
+    expect(useragent.browser).to eq('Wondery Crawler')
+  end
+
+  it "returns '#{version}' as its version" do
+    expect(useragent.version).to eq(version)
+  end
+
+  it { expect(useragent).to be_bot }
+  it { expect(useragent).not_to be_mobile }
+  it { expect(useragent).not_to be_desktop }
+  it { expect(useragent).not_to be_speaker }
+  it { expect(useragent.os).to be_nil }
+  it { expect(useragent.platform).to be_nil }
+end
+
 describe "UserAgent: wondery/android/v1.8.2/2314" do
   let(:useragent) { UserAgent.parse("wondery/android/v1.8.2/2314") }
 
@@ -47,4 +64,47 @@ describe "UserAgent: wondery/ios/v1.4.2/1 ,AppleCoreMedia/1.0.0.17G64 (iPhone; U
   let(:useragent) { UserAgent.parse("wondery/ios/v1.4.2/1 ,AppleCoreMedia/1.0.0.17G64 (iPhone; U; CPU OS 13_6 like Mac OS X; en_us)") }
 
   it_behaves_like 'Wondery', '1.4.2', 'iPhone', 'iOS 13.6'
+end
+
+describe 'UserAgent: wondery/develop' do
+  let(:useragent) { UserAgent.parse('wondery/develop') }
+
+  it_behaves_like 'Wondery Crawler', 'develop'
+end
+
+describe 'UserAgent: wondery/development' do
+  let(:useragent) { UserAgent.parse('wondery/development') }
+
+  it_behaves_like 'Wondery Crawler', 'development'
+end
+
+describe 'UserAgent: wondery/prod' do
+  let(:useragent) { UserAgent.parse('wondery/prod') }
+
+  it_behaves_like 'Wondery Crawler', 'prod'
+end
+
+describe 'UserAgent: wondery/production' do
+  let(:useragent) { UserAgent.parse('wondery/production') }
+
+  it_behaves_like 'Wondery Crawler', 'production'
+end
+
+describe 'UserAgent: wondery/staging' do
+  let(:useragent) { UserAgent.parse('wondery/staging') }
+
+  it_behaves_like 'Wondery Crawler', 'staging'
+end
+
+describe 'UserAgent: wondery/stage' do
+  let(:useragent) { UserAgent.parse('wondery/stage') }
+
+  it_behaves_like 'Wondery Crawler', 'stage'
+end
+
+describe 'UserAgent: wondery/not_crawler' do
+  let(:useragent) { UserAgent.parse('wondery/not_crawler') }
+
+  it { expect(useragent.browser).not_to eq('Wondery Crawler') }
+  it { expect(useragent).not_to be_bot }
 end